requirements.txt
streamlit<=1.24.0
pandas<=2.1.3
python-dotenv<=1.0.0
matplotlib<=3.8.0
numpy<=1.26.4
os<=0.1.2
fastapi<=0.116.2
OpenAI<=0.27.8